paperid

ID number of the paper on the journal website

volume

Volume number of the journal

issue

Issue number of the journal

url

Official website url of the paper

journal

Full name of the journal

title

Title of the paper

published_year

Year of publication

is_supp

Whether the paper has supplementary materials

num_supp

Number of supplementary material files

supp_file_type

File type of the supplementary materials

supp_url

Website url of the supplementary materials

num_authors

Number of the authors

first_author_name

Name of the first author

first_author_affiliation

Academic affiliation of the first author

first_author_affiliation_region

Country or region of the first author parsed from first_author_affiliation variable

first_author_email

Email of the first author

first_author_orcid

ORCID of the first author

correspondence_author_name

Name of the correspondence author

correspondence_author_affiliation

Academic affiliation of the correspondence author

correspondence_author_affiliation_region

Country or region of the correspondence author parsed from correspondence_author_affiliation variable

correspondence_author_email

Email of the correspondence author

correspondence_author_orcid

ORCID of the correspondence author

has_das

Whether the paper has a data availability statement

das

Original data availability statement of the paper

das_type

Type of the data availability statement #todo

das_repo_url

Website url of the data if the relevant data of the paper is shared on a public repository

keywords

List of keywords of the paper

paperid

ID number of the paper on the journal website

volume

Volume number of the journal

issue

Issue number of the journal

url

Official website url of the paper

journal

Full name of the journal

title

Title of the paper

published_year

Year of publication

is_supp

Whether the paper has supplementary materials

num_supp

Number of supplementary material files

supp_file_type

File type of the supplementary materials

supp_url

Website url of the supplementary materials

num_authors

Number of the authors

first_author_name

Name of the first author

first_author_affiliation

Academic affiliation of the first author

first_author_affiliation_region

Country or region of the first author parsed from first_author_affiliation variable

first_author_email

Email of the first author

first_author_orcid

ORCID of the first author

correspondence_author_name

Name of the correspondence author

correspondence_author_affiliation

Academic affiliation of the correspondence author

correspondence_author_affiliation_region

Country or region of the correspondence author parsed from correspondence_author_affiliation variable

correspondence_author_email

Email of the correspondence author

correspondence_author_orcid

ORCID of the correspondence author

has_das

Whether the paper has a data availability statement

das

Original data availability statement of the paper

das_type

Type of the data availability statement #todo

das_repo_url

Website url of the data if the relevant data of the paper is shared on a public repository

keywords

List of keywords of the paper

uncnewsletter

Format

An object of class tbl_df (inherits from tbl, data.frame) with 173 rows and 27 columns.